added new migration for french translations

+++ app/src/main/java/com/shid/mosquefinder/Data/database/entities/Category.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "category")
 data class Category(
-    @PrimaryKey(autoGenerate = true)  val id:Long,
+    @PrimaryKey(autoGenerate = true)  val id:Int,
     @ColumnInfo(name = "category_name") val categoryName: String,
 )
